+++ lib/ui/screens/tables/components/add_table_dialog.dart
@@ -113,11 +113,11 @@ class _AddTableDialogState extends ConsumerState<AddTableDialog> {
                         style: labelStyle,
                       ),
                       QtyEditor(
-                          qty: capacity,
+                          qty: capacity.toDouble(),
                           min: 1,
                           onChange: (value) {
                             setState(() {
-                              capacity = value;
+                              capacity = value.toInt();
                             });
                           }),
                     ],
